Add unit tests for print commands: YAZDIR and YAZDIRMAPROFİLİ
This commit introduces a comprehensive test suite for the print commands in KentOSCad, specifically focusing on the YAZDIR and YAZDIRMAPROFİLİ commands. The tests validate command execution, profile management, and error handling, ensuring that the command layer functions correctly without reliance on external dependencies like Qt. Key scenarios include profile listing, addition, deletion, and validation of command parameters, as well as ensuring that sensitive information is not logged. The tests also confirm that the command behaves consistently when invoked from both the command line and scripts.

+++ src/io/include/kentos_cad/io/service.hpp
@@ -44,12 +44,12 @@ namespace kentos::io {
 /// same numbers are what the wizard prints beside each tick box.
 struct ImportProbe
 {
-    std::string driver;                                        ///< DXF, DWG, GPKG…
-    std::string crs;                                           ///< what the file declared
-    std::uint64_t entities{0};                                 ///< across every layer
-    std::vector<std::pair<std::string, std::uint64_t>> layers; ///< name, entity count
-    ImportDiagnostics diagnostics;                             ///< the reader's own words, levelled
-    std::vector<VectorField> fields;                           ///< every attribute field, per layer
+    std::string driver;                                      ///< DXF, DWG, GPKG…
+    std::string crs;                                         ///< what the file declared
+    std::uint64_t entities{0};                               ///< across every layer
+    std::vector<std::pair<std::string, std::size_t>> layers; ///< name, entity count
+    ImportDiagnostics diagnostics;                           ///< the reader's own words, levelled
+    std::vector<VectorField> fields;                         ///< every attribute field, per layer
 };
 
 /// Reads `path` into `scratch` — a document the caller owns and the user has never
